Update .env
Adding SOP_TYPE_ID=1544 as a placeholder, even though the logic to use this hasn't been written yet. 1544 is the default "Approved Software" category, but may be swapped out for 1547 (Approved Software: Dimensional) or others.  By adding this now, we avoid the risk of mistyping SOP_TYPE_ID in future repos created from this template.